fix: Use f64 for distance throughout, fix UREVO missing FTMS distance
Change total_distance_m from i32 to f64 across server, client, and walker
for sub-meter precision. The UREVO treadmill reports speed via BLE but not
cumulative distance, so the walker now computes distance from speed × elapsed
time when FTMS distance is unavailable.

Also fixes: debug_walk handler only sets speed/flags (tick computes delta),
gold awarded per 10m milestone, combat events dismissed when no route set,
blocking events log at info level.

+++ crates/gameclient/src/supabase.rs
@@ -31,7 +31,7 @@ pub struct PlayerRow {
     pub id: String,
     pub name: String,
     pub current_speed_kmh: f32,
-    pub total_distance_m: i32,
+    pub total_distance_m: f64,
     pub is_walking: bool,
     pub map_tile_x: Option<i32>,
     pub map_tile_y: Option<i32>,
